requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dlib @ file:///D:/Repository/penelitian_2024/prototype_face_recognition/dlib-19.24.1-cp311-cp311-win_amd64.whl#sha256=6f1a5ee167975d7952b28e0ce4495f1d9a77644761cf5720fb66d7c6188ae496
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-MySQL==1.5.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.1
opencv-python==4.9.0.80
pillow==10.3.0
PyMySQL==1.1.1
Werkzeug==3.0.3
